Migrate salt.utils.thin unit tests to pytest
